Organize code from main.rb into tic_tac_toe.rb
