+++ helpers/get_human_input.rb
