+++ helpers/minmax.rb
